+++ AddOns/LiteButtonAuras/CHANGELOG.md
@@ -1,6 +1,13 @@
 # LiteButtonAuras
 
-## [10.2.6-2](https://github.com/xod-wow/LiteButtonAuras/tree/10.2.6-2) (2024-03-22)
-[Full Changelog](https://github.com/xod-wow/LiteButtonAuras/compare/10.2.6-1...10.2.6-2) [Previous Releases](https://github.com/xod-wow/LiteButtonAuras/releases)
+## [10.2.6-3](https://github.com/xod-wow/LiteButtonAuras/tree/10.2.6-3) (2024-05-01)
+[Full Changelog](https://github.com/xod-wow/LiteButtonAuras/compare/10.2.6-2...10.2.6-3) [Previous Releases](https://github.com/xod-wow/LiteButtonAuras/releases)
 
-- Fix error with macro items that don't have spells  
+- Update ToC for Cataclysm Classic  
+- Update classic\_era ToC  
+- Make the ignored spells only ignore the base name matching  
+- Fix some bugs with spells by name stuff  
+- Support (but not yet allow) AuraMap by name only  
+- Move AuraMap into Overlay where it probably always belonged  
+- I dunno what I'm doing with spell inputs  
+- Working on spell ID input widget